notebooks/requirements-base.txt
jq
jp
wget

python-dotenv

jsonpath-ng
gql[all]

pandas
dask[complete]
psycopg2-binary
sqlalchemy
sqlacodegen-v2

PyGithub

opentelemetry-api
opentelemetry-sdk

